Moving function for coalescing network codes to base class
